pyproject.toml
[tool.poetry]
name = "deethon"
version = "0.3.3"
description = "Python3 library to easily download music from Deezer"
authors = [
    "Aykut Yilmaz <aykuxt@gmail.com>"
]
license = "GPL-3.0-or-later"
readme = "README.md"
repository = "https://github.com/deethon/deethon"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Operating System :: OS Independent",
]

[tool.poetry.dependencies]
python = "^3.8"
mutagen = "^1.44.0"
requests = "^2.23.0"
pycryptodome = "^3.9.7"

[tool.poetry.dev-dependencies]
pytest = "^5.4.2"
yapf = "^0.30.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.dephell.main]
from = {format = "poetry", path = "pyproject.toml"}
envs = ["main"]
to = {format = "setuppy", path = "setup.py"}
